refactor: viewModel 코드 리팩토링
