+++ src/components/Cart/CartList.tsx
@@ -1,5 +1,5 @@
 import getCartItems from '../../api/getCartItems';
-import createCartItemsViewModel, { CartItemViewModel } from '../../api/model/createCartItemsViewModel';
+import { createCartItemViewModel, CartItemViewModel } from '../../api/model/createCartItemViewModel';
 import patchCartItem from '../../api/patchCartItem';
 import { useApiContext } from '../../contexts/ApiContext';
 import { useErrorContext } from '../../contexts/ErrorContext';
@@ -19,42 +19,57 @@ export default function CartList({
   const { showError } = useErrorContext();
   const { fetcher: refetchCart } = useApiContext({ fetchFn: getCartItems, key: 'getCartItems' });
 
-  const viewModel = createCartItemsViewModel(cartItems);
-
   const handleMinus = async (item: CartItemViewModel) => {
-    await patchCartItem(item.id, item.cartQuantity - 1);
-    await refetchCart();
+    try {
+      await patchCartItem(item.id, item.cartQuantity - 1);
+      await refetchCart();
+    } catch (err) {
+      if (err instanceof Error) {
+        showError(err);
+      }
+    }
   };
+
   const handlePlus = async (item: CartItemViewModel) => {
     if (item.cartQuantity >= item.productQuantity) {
       showError(new Error('수량을 초과해서 담을 수 없어요.'));
+      return;
     }
-    await patchCartItem(item.id, item.cartQuantity + 1);
-    await refetchCart();
+    try {
+      await patchCartItem(item.id, item.cartQuantity + 1);
+      await refetchCart();
+    } catch (err) {
+      if (err instanceof Error) {
+        showError(err);
+      }
+    }
   };
 
   return (
     <>
-      {viewModel.map((item) => (
-        <div key={item.id} css={styles.cartItemWrapper}>
-          <div css={styles.cartItem}>
-            <div css={styles.cartImageWrapper}>
-              <Image src={item.imageUrl} alt={`${item.title} 상품 이미지`} />
-            </div>
-            <div css={styles.cartTextBlock}>
-              <h3 css={styles.titleCss}>{item.title}</h3>
-              <p>{item.price}</p>
-              <Counter
-                value={item.cartQuantity}
-                onDecrement={() => handleMinus(item)}
-                onIncrement={() => handlePlus(item)}
-              />
-            </div>
-          </div>
+      {cartItems.map((item) => {
+        const viewModel = createCartItemViewModel(item);
 
-          <RemoveFromCartButton onClick={() => onClick(item)} />
-        </div>
-      ))}
+        return (
+          <div key={viewModel.id} css={styles.cartItemWrapper}>
+            <div css={styles.cartItem}>
+              <div css={styles.cartImageWrapper}>
+                <Image src={viewModel.imageUrl} alt={`${viewModel.title} 상품 이미지`} />
+              </div>
+              <div css={styles.cartTextBlock}>
+                <h3 css={styles.titleCss}>{viewModel.title}</h3>
+                <p>{viewModel.price}</p>
+                <Counter
+                  value={viewModel.cartQuantity}
+                  onDecrement={() => handleMinus(viewModel)}
+                  onIncrement={() => handlePlus(viewModel)}
+                />
+              </div>
+            </div>
+            <RemoveFromCartButton onClick={() => onClick(viewModel)} />
+          </div>
+        );
+      })}
     </>
   );
 }
